# docs/openapi/openapi-generated.yaml
openapi: 3.0.3
info:
  title: Photo Editor API
  version: 1.0.0
  description: API contracts generated from Zod schemas for the Photo Editor application
  license:
    name: MIT
  contact:
    name: API Support
servers:
  - url: http://localhost:4566
    description: LocalStack development server
  - url: https://api.photoeditor.dev
    description: Development environment
  - url: https://api.photoeditor.com
    description: Production environment
paths:
...
  /v1/jobs/{id}:
    get:
      operationId: getJobStatus
      summary: Get job status
      description: Retrieves the current processing status and metadata for a job by its ID.
      tags:
        - Jobs
      parameters:
        - name: id
          in: path
          description: Job UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum:
                      - QUEUED
                      - PROCESSING
                      - EDITING
                      - COMPLETED
                      - FAILED
                  createdAt:
                    type: string
                    format: date-time
                  updatedAt:
                    type: string
                    format: date-time
                  error:
                    type: string
                required:
                  - jobId
                  - status
                  - createdAt
                  - updatedAt
                additionalProperties: false
        '400':
          description: Bad Request - Invalid input or validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  timestamp:
                    type: string
                    format: date-time
                  requestId:
                    type: string
                required:
                  - error
                  - timestamp
                  - requestId
                additionalProperties: false
        '404':
          description: Not Found - Resource does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  timestamp:
                    type: string
                    format: date-time
                  requestId:
                    type: string
                required:
                  - error
                  - timestamp
                  - requestId
                additionalProperties: false
        '500':
          description: Internal Server Error - Unexpected server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  timestamp:
                    type: string
                    format: date-time
                  requestId:
                    type: string
                required:
                  - error
                  - timestamp
                  - requestId
                additionalProperties: false
...
  /v1/jobs/{id}/download:
    get:
      operationId: downloadJobResult
      summary: Get download URL for completed job
      description: Generates a presigned S3 download URL for a completed job's edited photo. Returns 400 if job is not completed.
      tags:
        - Jobs
        - Download
      parameters:
        - name: id
          in: path
          description: Job UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  downloadUrl:
                    type: string
                    format: uri
                  expiresAt:
                    type: string
                    format: date-time
                  jobId:
                    type: string
                    format: uuid
                  status:
                    type: string
                required:
                  - downloadUrl
                  - expiresAt
                  - jobId
                  - status
                additionalProperties: false
        '400':
          description: Bad Request - Invalid input or validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  timestamp:
                    type: string
                    format: date-time
                  requestId:
                    type: string
                required:
                  - error
                  - timestamp
                  - requestId
                additionalProperties: false
        '404':
          description: Not Found - Resource does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  timestamp:
                    type: string
                    format: date-time
                  requestId:
                    type: string
                required:
                  - error
                  - timestamp
                  - requestId
                additionalProperties: false
        '500':
          description: Internal Server Error - Unexpected server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  timestamp:
                    type: string
                    format: date-time
                  requestId:
                    type: string
                required:
                  - error
                  - timestamp
                  - requestId
                additionalProperties: false
...